+++ changelog.md
@@ -1,3 +1,7 @@
+### 1.0.4 12/06/24
+Add images for each button
+### 1.0.33 11/10/24
+Prevent touch controls from being created twice
 ### 1.0.32 9/30/24
 Moved controls up to prevent accidentally leaving level
 ### 1.0.31 9/29/24